✨ feat(UI, form): 사용자 경험 및 폼 상태 관리 개선
변경한 이유:
- 사용자 경험을 향상시키고 폼 상태 관리의 일관성을 확보하기 위함.
- 마이페이지 UI, 게시글 생성 폼 초기화, 헤더 겹침 문제 등 여러 UX/UI 관련 불편 사항을 해결.
변경 내용:
- 마이페이지에서 이름과 대표 가게 항목 간의 간격을 조정하여 가독성 개선
- 이름 수정 완료 시 자동으로 이전 페이지로 돌아가도록 라우팅 로직 추가
- 게시글 생성 성공 후 다시 폼으로 진입 시 currentStepIndex를 1단계(StepOne)로 초기화하고 폼 데이터를 리셋하도록 수정
- 기록 보기 및 기록 상세보기 페이지에서 헤더가 겹쳐 보이던 문제를 화면 vh 조정으로 해결
